Move test to venice-common folder
