+++ services/venice-controller/src/test/java/com/linkedin/venice/controller/TestVeniceParentHelixAdmin.java
@@ -58,7 +58,6 @@ import com.linkedin.venice.meta.HybridStoreConfigImpl;
 import com.linkedin.venice.meta.OfflinePushStrategy;
 import com.linkedin.venice.meta.PersistenceType;
 import com.linkedin.venice.meta.ReadStrategy;
-import com.linkedin.venice.meta.RegionPushDetails;
 import com.linkedin.venice.meta.RoutingStrategy;
 import com.linkedin.venice.meta.Store;
 import com.linkedin.venice.meta.StoreInfo;
@@ -68,9 +67,6 @@ import com.linkedin.venice.meta.ZKStore;
 import com.linkedin.venice.offsets.OffsetRecord;
 import com.linkedin.venice.partitioner.InvalidKeySchemaPartitioner;
 import com.linkedin.venice.pushmonitor.ExecutionStatus;
-import com.linkedin.venice.pushmonitor.OfflinePushStatus;
-import com.linkedin.venice.pushmonitor.PartitionStatus;
-import com.linkedin.venice.pushmonitor.StatusSnapshot;
 import com.linkedin.venice.schema.avro.DirectionalSchemaCompatibilityType;
 import com.linkedin.venice.serialization.avro.AvroProtocolDefinition;
 import com.linkedin.venice.utils.DataProviderUtils;
@@ -82,7 +78,6 @@ import com.linkedin.venice.utils.Time;
 import com.linkedin.venice.utils.Utils;
 import com.linkedin.venice.utils.concurrent.VeniceConcurrentHashMap;
 import com.linkedin.venice.writer.VeniceWriter;
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
@@ -94,7 +89,6 @@ import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
-import org.apache.commons.lang.StringUtils;
 import org.apache.http.HttpStatus;
 import org.apache.kafka.clients.producer.RecordMetadata;
 import org.apache.kafka.common.TopicPartition;
@@ -2499,46 +2493,5 @@ public class TestVeniceParentHelixAdmin extends AbstractTestVeniceParentHelixAdm
     parentAdmin.createStore(clusterName, storeName, owner, keySchemaStr, valueSchemaStr);
     doReturn(clusterLockManager).when(resources).getClusterLockManager();
     verify(clusterLockManager).createClusterReadLock();
-  }
-
-  @Test
-  public void testDataRecoveryAPIs() {
-    final String storeName = "test";
-    final String owner = "test";
-    final int numOfPartition = 5;
-    final int replicationFactor = 3;
-    final String kafkaTopic = "test_v1";
-
-    OfflinePushStatus status = new OfflinePushStatus(
-        kafkaTopic,
-        numOfPartition,
-        replicationFactor,
-        OfflinePushStrategy.WAIT_N_MINUS_ONE_REPLCIA_PER_PARTITION);
-    LocalDateTime now = LocalDateTime.now();
-
-    for (int i = 0; i < numOfPartition; i++) {
-      PartitionStatus partition = new PartitionStatus(i);
-      for (int j = 0; j < replicationFactor; j++) {
-        partition.updateReplicaStatus("instanceId-" + j, ExecutionStatus.STARTED, StringUtils.EMPTY);
-        partition.updateReplicaStatus("instanceId-" + j, ExecutionStatus.COMPLETED, StringUtils.EMPTY);
-      }
-      status.setPartitionStatus(partition);
-    }
-
-    status.getStatusHistory().add(new StatusSnapshot(ExecutionStatus.STARTED, now.toString()));
-    status.getStatusHistory().add(new StatusSnapshot(ExecutionStatus.COMPLETED, now.plusHours(1).toString()));
-    doReturn(status).when(internalAdmin).retrievePushStatus(anyString(), anyString());
-
-    Store s = TestUtils.createTestStore(storeName, owner, System.currentTimeMillis());
-    when(internalAdmin.getRegionPushDetails(anyString(), anyString(), anyBoolean())).thenCallRealMethod();
-    doReturn(s).when(internalAdmin).getStore(anyString(), anyString());
-
-    RegionPushDetails details = internalAdmin.getRegionPushDetails(clusterName, storeName, true);
-    Assert.assertEquals(details.getPushEndTimestamp(), now.plusHours(1).toString());
-    Assert.assertEquals(details.getCurrentVersion().intValue(), Store.NON_EXISTING_VERSION);
-    Assert.assertEquals(details.getPartitionDetails().size(), numOfPartition);
-    for (int i = 0; i < numOfPartition; i++) {
-      Assert.assertEquals(details.getPartitionDetails().get(i).getReplicaDetails().size(), replicationFactor);
-    }
   }
 }
